# Improve speadLight and updateLighting performance