To test this, a superset of these changes and timing code is:
https://github.com/gregorycu/minetest/tree/updateLightingPerfWithStats

The observed results on Windows x64 Visual Studio build. 

Case 1: Updating an 80x80x80 area (500k blocks)

Using mod from here: https://github.com/paramat/testspreadlight

0.1% Meselamp
0.33% stone
0.66% air

Results are day and night.

Old: 632 ms 399 ms
New: 372 ms 230 ms
Difference: -41% -42%

Case 2: Spawning a jungle tree from the moretrees mod

Results are day and night (total for spawn_ltree)

Old: 277 ms 6 ms (297 ms)
New: 176 ms 5ms (193 ms)
Difference: -36% -16% (-35%)