@nerzhul 2016-01-03
please fix build @gregorycu