@paramat 2016-01-04
:+1: Tests seem good but i'm unqualified to review the code.