@paramat 2016-01-06
From IRC: <sapier> I can't proof this but imho prereserving should be the most important factor in this performance boost. if this is true size of prereservation would be  a relevant factor in it too