@gregorycu 2016-01-06
Maybe 100 should be replaced by 6x the number of nodes currently getting
processed. (6 is the spread factor).
On 07/01/2016 9:43 am, "Paramat" <notifications@github.com> wrote:

> From IRC: I can't proof this but imho prereserving should be the most
> important factor in this performance boost. if this is true size of
> prereservation would be a relevant factor in it too
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/minetest/minetest/pull/3523#issuecomment-169486015>.
>