@Fixer-007 2016-02-01
I will definitely test it when camera mod will be published (for benchmarking). I'm currently playing with those PRs.

If developers are sure that it boosts performance, then :+1 and merge anyway!