@kilbith 2016-02-16
@gregorycu caused a meshgen regression due to using a slower stdlib on Windows, so maybe we have to be very circonspect about this (`reserve` is used here).

http://irc.minetest.ru/minetest-dev/2016-02-16#i_4535233
http://irc.minetest.ru/minetest-dev/2016-02-16#i_4535236