@gregorycu 2016-02-16
:/ I doubt it.
On 16/02/2016 9:13 pm, "kilbith" <notifications@github.com> wrote:

> @gregorycu <https://github.com/gregorycu> caused a meshgen regression due
> to using a slower stdlib on Windows, so maybe we have to be very
> circonspect about this (reserve is used here).
>
> http://irc.minetest.ru/minetest-dev/2016-02-16#i_4535233
> http://irc.minetest.ru/minetest-dev/2016-02-16#i_4535236
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/minetest/minetest/pull/3523#issuecomment-184609192>.
>