@RealBadAngel 2016-02-16
@gregorycu in case of mesh collector, using two times reserve caused append() execution time ~230000 nanosecond instead of ~250ns