@nerzhul 2016-02-18
reserve on a vector should only be used at first vector initialization, not when using it because it reallocate the WHOLE memory